Split handlers into separate files
